requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
django-cors-headers==4.9.0
django_celery_results==2.6.0
djangorestframework==3.17.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.1
flatbuffers==25.12.19
idna==3.11
ImageIO==2.37.3
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lazy-loader==0.5
llvmlite==0.47.0
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
pillow==12.2.0
platformdirs==4.9.4
pooch==1.9.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psycopg2-binary==2.9.11
PyMatting==1.1.15
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rembg==2.0.74
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tifffile==2026.3.3
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
